# backend/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Tchat API - Telegram SEA Edition
  description: |
    Comprehensive API documentation for Tchat (Telegram SEA Edition), a Southeast Asian messaging and commerce platform.

    ## Features
    - **Authentication**: OTP-based authentication with JWT tokens
    - **Messaging**: Real-time messaging with WebSocket support
    - **Payments**: Multi-currency wallet system with regional support
    - **Commerce**: E-commerce platform with shop and product management
    - **Notifications**: Multi-channel notification system with templates

    ## Regional Support
    - **Countries**: Thailand, Singapore, Indonesia, Malaysia, Philippines, Vietnam
    - **Currencies**: THB, SGD, IDR, MYR, PHP, VND, USD
    - **Languages**: English, Thai, Bahasa Indonesia, Bahasa Malaysia, Filipino, Vietnamese

    ## Security
    - JWT-based authentication with refresh tokens
    - Rate limiting on all endpoints
    - CORS support for web applications
    - Webhook signature verification

    ## WebSocket Support
    Real-time messaging is available via WebSocket connections at `/ws/messaging`
  version: "1.0.0"
  contact:
    name: Tchat API Support
    url: https://tchat.sea/support
    email: api-support@tchat.sea
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.tchat.sea/v1
    description: Production server
  - url: https://staging-api.tchat.sea/v1
    description: Staging server
  - url: http://localhost:8080/api/v1
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from authentication endpoints
  schemas:
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
          example: "Invalid request payload"
        details:
          type: string
          description: Additional error details
          example: "Field 'email' is required"
        code:
          type: string
          description: Error code for programmatic handling
          example: "VALIDATION_ERROR"
...
    VerifyOTPRequest:
      type: object
      required:
        - phone_number
        - country_code
        - otp_code
      properties:
        phone_number:
          type: string
          pattern: '^[0-9]{8,15}$'
          example: "812345678"
        country_code:
          type: string
          enum: [TH, SG, ID, MY, PH, VN]
          example: "TH"
        otp_code:
          type: string
          pattern: '^[0-9]{6}$'
          description: 6-digit OTP code
          example: "123456"
        device_info:
          type: object
          description: Device information for security
          properties:
            device_id:
              type: string
              example: "mobile_app_ios_abc123"
            platform:
              type: string
              enum: [web, mobile_ios, mobile_android]
              example: "mobile_ios"
            app_version:
              type: string
              example: "1.0.0"
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token (expires in 1 hour)
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refresh_token:
          type: string
          description: Refresh token (expires in 30 days)
          example: "rt_abc123def456..."
        expires_in:
          type: integer
          description: Access token expiration time in seconds
          example: 3600
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        phone_number:
          type: string
          example: "+66812345678"
        country_code:
          type: string
          example: "TH"
        status:
          type: string
          enum: [active, inactive, suspended]
          example: "active"
        profile:
          type: object
          properties:
            display_name:
              type: string
              example: "John Doe"
            avatar_url:
              type: string
              format: uri
              example: "https://cdn.tchat.sea/avatars/user123.jpg"
            locale:
              type: string
              example: "en"
        created_at:
          type: string
          format: date-time
          example: "2023-12-01T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2023-12-01T10:00:00Z"

    # Messaging schemas
...
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Authentication required"
            code: "UNAUTHORIZED"
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Validation failed"
            details: "Field 'email' is required"
            code: "VALIDATION_ERROR"
    RateLimitError:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Rate limit exceeded"
            code: "RATE_LIMIT_EXCEEDED"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal server error"
            code: "INTERNAL_ERROR"
paths:
...
  /auth/otp/verify:
    post:
      tags:
        - Authentication
      summary: Verify OTP and authenticate
      description: Verify the OTP code and return authentication tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyOTPRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          description: Invalid OTP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Invalid or expired OTP"
                code: "INVALID_OTP"
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Use refresh token to get a new access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
                  example: "rt_abc123def456..."
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  expires_in:
                    type: integer
                    example: 3600
        '401':
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Invalid or expired refresh token"
                code: "INVALID_REFRESH_TOKEN"
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      description: Invalidate the current session and tokens
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Logout successful"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Messaging endpoints
...